pyproject.toml
[project]
name = "openvassal"
version = "0.2.0"
description = "OpenVassal — Personal knowledge base with multi-agent orchestration (CrewAI + mem0)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "crewai[tools]>=0.100.1",
    "mem0ai>=0.1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "sqlite-utils>=3.35",
    "python-dotenv>=1.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
openvassal = "openvassal.main:cli"

[tool.setuptools.packages.find]
include = ["openvassal*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
